Move files to commands/*
