+++ caldir-provider-google/src/google/mod.rs
@@ -1,4 +1,3 @@
 pub mod actions;
-pub mod api;
 pub mod from_google;
 pub mod to_google;
