+++ caldir-provider-google/src/main.rs
@@ -7,12 +7,13 @@
 //!   ~/.config/caldir/providers/google/credentials.json
 //!   ~/.config/caldir/providers/google/tokens/{account}.json
 
+mod commands;
 mod config;
 mod google;
 mod types;
 
+use anyhow::Result;
 use caldir_core::protocol::{Command, Request, Response};
-use serde::Deserialize;
 use std::io::{self, BufRead, Write};
 
 /// Google's alias for the user's main calendar
@@ -48,151 +49,31 @@ async fn main() {
             }
         };
 
-        let response = handle_request(request).await;
+        let result = handle_request(request).await;
 
-        if writeln!(stdout, "{}", response).is_err() || stdout.flush().is_err() {
+        let protocol_response = match result {
+            Ok(data) => Response::success(data),
+            Err(e) => Response::error(&format!("Error handling request: {:#}", e)),
+        };
+
+        if writeln!(stdout, "{}", protocol_response).is_err() || stdout.flush().is_err() {
             break; // Parent likely killed us, exit silently
         }
     }
 }
 
-async fn handle_request(request: Request) -> String {
-    match request.command {
-        Command::Authenticate => handle_authenticate().await,
-        Command::ListCalendars => handle_list_calendars(&request.params).await,
-        Command::ListEvents => handle_list_events(&request.params).await,
-        Command::CreateEvent => handle_create_event(&request.params).await,
-        Command::UpdateEvent => handle_update_event(&request.params).await,
-        Command::DeleteEvent => handle_delete_event(&request.params).await,
-    }
-}
-
-async fn handle_authenticate() -> String {
-    match google::actions::authenticate().await {
-        Ok(account) => Response::success(account),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct FetchCalendarsParams {
-    google_account: String,
-}
-
-async fn handle_list_calendars(params: &serde_json::Value) -> String {
-    let params: FetchCalendarsParams = match serde_json::from_value(params.clone()) {
-        Ok(p) => p,
-        Err(e) => return Response::error(&format!("Invalid params: {}", e)),
+async fn handle_request(request: Request) -> Result<String> {
+    let result = match request.command {
+        Command::Authenticate => commands::handle_authenticate().await,
+        Command::ListCalendars => commands::handle_list_calendars(&request.params).await,
+        Command::ListEvents => commands::handle_list_events(&request.params).await,
+        Command::CreateEvent => commands::handle_create_event(&request.params).await,
+        Command::UpdateEvent => {
+            commands::handle_update_event(&client, &calendar_id, &request.params)
+                .await,
+        }
+        Command::DeleteEvent => commands::handle_delete_event(&request.params).await,
     };
 
-    match google::api::fetch_calendars(&params.google_account).await {
-        Ok(calendars) => Response::success(calendars),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct FetchEventsParams {
-    google_account: String,
-    google_calendar_id: Option<String>,
-    #[serde(default)]
-    time_min: Option<String>,
-    #[serde(default)]
-    time_max: Option<String>,
-}
-
-async fn handle_list_events(params: &serde_json::Value) -> String {
-    let params: FetchEventsParams = match serde_json::from_value(params.clone()) {
-        Ok(p) => p,
-        Err(e) => return Response::error(&format!("Invalid params: {}", e)),
-    };
-
-    let calendar_id = params
-        .google_calendar_id
-        .as_deref()
-        .unwrap_or(DEFAULT_CALENDAR_ID);
-
-    match google::api::fetch_events(
-        &params.google_account,
-        calendar_id,
-        params.time_min.as_deref(),
-        params.time_max.as_deref(),
-    )
-    .await
-    {
-        Ok(events) => Response::success(events),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct CreateEventParams {
-    google_account: String,
-    google_calendar_id: Option<String>,
-    event: types::Event,
-}
-
-async fn handle_create_event(params: &serde_json::Value) -> String {
-    let params: CreateEventParams = match serde_json::from_value(params.clone()) {
-        Ok(p) => p,
-        Err(e) => return Response::error(&format!("Invalid params: {}", e)),
-    };
-
-    let calendar_id = params
-        .google_calendar_id
-        .as_deref()
-        .unwrap_or(DEFAULT_CALENDAR_ID);
-
-    match google::api::create_event(&params.google_account, calendar_id, &params.event).await {
-        Ok(event) => Response::success(event),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct UpdateEventParams {
-    google_account: String,
-    google_calendar_id: Option<String>,
-    event: types::Event,
-}
-
-async fn handle_update_event(params: &serde_json::Value) -> String {
-    let params: UpdateEventParams = match serde_json::from_value(params.clone()) {
-        Ok(p) => p,
-        Err(e) => return Response::error(&format!("Invalid params: {}", e)),
-    };
-
-    let calendar_id = params
-        .google_calendar_id
-        .as_deref()
-        .unwrap_or(DEFAULT_CALENDAR_ID);
-
-    match google::api::update_event(&params.google_account, calendar_id, &params.event).await {
-        Ok(event) => Response::success(event),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct DeleteEventParams {
-    google_account: String,
-    google_calendar_id: Option<String>,
-    event_id: String,
-}
-
-async fn handle_delete_event(params: &serde_json::Value) -> String {
-    let params: DeleteEventParams = match serde_json::from_value(params.clone()) {
-        Ok(p) => p,
-        Err(e) => return Response::error(&format!("Invalid params: {}", e)),
-    };
-
-    let calendar_id = params
-        .google_calendar_id
-        .as_deref()
-        .unwrap_or(DEFAULT_CALENDAR_ID);
-
-    match google::api::delete_event(&params.google_account, calendar_id, &params.event_id).await {
-        Ok(()) => Response::success(()),
-        Err(e) => Response::error(&format!("{:#}", e)),
-    }
+    result
 }
